requirements.txt
# Core dependencies for anki-mcp-server
fastmcp>=0.4.2
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
uvicorn>=0.22.0
